Address PR feedback: remove redundant duration parser, use i32 for limit/offset
- Remove parse_duration_to_millis() and its 8 tests from util.rs (redundant with
  parse_time_to_unix_millis regex logic; --interval now takes raw milliseconds)
- Change --limit and --offset from i64 to i32 for consistency with other commands
  (logs, events, etc. all use i32)
- Update --interval to Option<i64> (raw milliseconds) matching the API's
  default_interval field directly

+++ src/commands/ddsql.rs
@@ -13,8 +13,8 @@ fn build_request(
     from: &str,
     to: &str,
     interval: Option<i64>,
-    limit: Option<i64>,
-    offset: Option<i64>,
+    limit: Option<i32>,
+    offset: Option<i32>,
 ) -> Result<Value> {
     let default_start =
         util::parse_time_to_unix_millis(from).map_err(|e| anyhow!("invalid --from: {e}"))?;
@@ -52,8 +52,8 @@ pub async fn table(
     from: &str,
     to: &str,
     interval: Option<i64>,
-    limit: Option<i64>,
-    offset: Option<i64>,
+    limit: Option<i32>,
+    offset: Option<i32>,
 ) -> Result<()> {
     let body = build_request(
         "ddsql_table_request",
@@ -74,8 +74,8 @@ pub async fn csv(
     from: &str,
     to: &str,
     interval: Option<i64>,
-    limit: Option<i64>,
-    offset: Option<i64>,
+    limit: Option<i32>,
+    offset: Option<i32>,
 ) -> Result<()> {
     let body = build_request(
         "ddsql_table_request",
